freezing testing while fixing android stuff
