+++ e2e/ios/.gitignore
@@ -1,0 +1,1 @@
+desired_caps.py
