+++ e2e/ios/desired_caps_sample.py
@@ -1,0 +1,5 @@
+desired_caps = {}
+desired_caps['platformName'] = 'iOS'
+desired_caps['platformVersion'] = '11.1'
+desired_caps['deviceName'] = 'iPhone 7'
+desired_caps['app'] = 'https://s3.amazonaws.com/bookit-mobile-artifacts/local-testing.ipa'
