+++ src/screens/HomeScreen.js
@@ -6,7 +6,7 @@ import {
   FlatList,
 } from 'react-native'
 import { connect } from 'react-redux'
-import { DateTime, Duration } from 'luxon'
+import { DateTime, Duration, Info } from 'luxon'
 
 import {
   fetchLocations,
@@ -136,9 +136,9 @@ class HomeScreen extends React.Component {
 
 const mapStateToProps = (state) => {
   const { selectedLocation, locations, bookables } = state
-
   // Set Booking defaults
-  const start = DateTime.local().plus({ hour: 1 }).setZone(selectedLocation.timeZone)
+  const start = DateTime.local().plus({ hour: 1 })
+  console.log(' lol wtf???', Info.features())
   const bookingDuration = Duration.fromObject({ minutes: 30 })
 
   return ({
